connexion-example/requirements.txt
connexion[swagger-ui]==2.13.1
Flask-Sqlalchemy==2.5.1
bcrypt==3.2.2
pyjwt==2.0.1
